+++ src/NASSA/Types.hs
@@ -10,12 +10,13 @@ import           Control.Monad       (guard, mzero)
 import           Data.Aeson          (FromJSON, ToJSON (..), Value (String),
                                       parseJSON, withObject, withText, (.:),
                                       (.:?))
-import           Data.Char           (digitToInt)
+import           Data.Char           (digitToInt, toLower)
 import           Data.List           (intercalate)
 import qualified Data.Text           as T
 import qualified Data.Text.Encoding  as TS
 import           Data.Time           (Day)
 import           Data.Version        (Version, makeVersion, showVersion)
+import           System.FilePath     (takeExtension)
 import qualified Text.Email.Validate as TEV
 import qualified Text.Parsec         as P
 import qualified Text.Parsec.Text    as P
@@ -31,7 +32,8 @@ data NassaModuleYamlStruct = NassaModuleYamlStruct {
     , _nassaYamlModuleVersion       :: Version
     , _nassaYamlContributors        :: [Contributor]
     , _nassaYamlLastUpdateDate      :: Day
-    , _nassaYamlDescription         :: String
+    , _nassaYamlDescription         :: ModuleDescription
+    , _nassaYamlCoverImagePath      :: Maybe ModuleImagePath
     , _nassaYamlRelatedModules      :: Maybe [ModuleID]
     , _nassaYamlReferences          :: Maybe ReferenceStruct
     , _nassaYamlDomainKeywords      :: Maybe DomainKeyword
@@ -55,6 +57,7 @@ instance FromJSON NassaModuleYamlStruct where
         <*> v .:  "contributors"
         <*> v .:  "lastUpdateDate"
         <*> v .:  "description"
+        <*> v .:? "coverImage"
         <*> v .:? "relatedModules"
         <*> v .:? "references"
         <*> v .:? "domainKeywords"
@@ -97,9 +100,36 @@ instance Show ModuleTitle where
 
 instance FromJSON ModuleTitle where
     parseJSON (String s) =
-        if T.length s <= 100
+        if T.length s <= 50
         then pure $ ModuleTitle $ T.unpack s
-        else fail "module title must not be longer than 100 characters"
+        else fail "module title must not be longer than 50 characters"
+    parseJSON _ = mzero
+
+newtype ModuleDescription = ModuleDescription String
+    deriving (Eq)
+
+instance Show ModuleDescription where
+    show (ModuleDescription s) = s
+
+instance FromJSON ModuleDescription where
+    parseJSON (String s) =
+        if T.length s <= 300
+        then pure $ ModuleDescription $ T.unpack s
+        else fail "module description must not be longer than 300 characters"
+    parseJSON _ = mzero
+
+newtype ModuleImagePath = ModuleImagePath FilePath
+    deriving (Eq)
+
+instance Show ModuleImagePath where
+    show (ModuleImagePath p) = p
+
+instance FromJSON ModuleImagePath where
+    parseJSON (String s) =
+        let path = T.unpack s
+        in if map toLower (takeExtension path) `elem` ["png", "jpg", "jpeg", "svg"]
+           then pure $ ModuleImagePath $ path
+           else fail "cover image must have extension png, jpg, jpeg, or svg"
     parseJSON _ = mzero
 
 data ModuleType =